Enum tame_oauth::gcp::TokenOrRequest

pub enum TokenOrRequest {
    Token(Token),
    Request {
        request: Request<Vec<u8>>,
        reason: RequestReason,
        scope_hash: u64,
    },
}

Either a valid token, or an HTTP request that can be used to acquire one

Variants

Token(Token)

A valid token that can be supplied in an API request

Request

Fields of Request

request: Request<Vec<u8>>

The parts of an HTTP request that must be sent to acquire the token, in the client of your choice

reason: RequestReason

The reason we need to retrieve a new token

scope_hash: u64

An opaque hash of the scope(s) for which the request was constructed